optimized best_path_trans svm computations.

+++ src/kernel/CommWordKernel.h
@@ -70,10 +70,10 @@ class CCommWordKernel: public CSimpleKernel<WORD>
 			  return value/sqrtdiag_lhs[seq_num];
 			  break;
 		  case SQRTLEN_NORMALIZATION:
-			  return value/sqrt(sqrt(seq_len));
+			  return value/sqrt(sqrt((double)seq_len));
 			  break;
 		  case LEN_NORMALIZATION:
-			  return value/sqrt(seq_len);
+			  return value/sqrt((double)seq_len);
 			  break;
 		  case SQLEN_NORMALIZATION:
 			  return value/seq_len;
